Minor modif to the unchecked-return-transfer - Inherits from Unused return (avoid duplicate code) - Rename unchecked-return - Update description/recommendations
Additionally removed txt_description from all the UnusedReturn derived
class (its not used anymore)
