+++ slither/detectors/operations/unchecked_low_level_return_values.py
@@ -34,7 +34,5 @@ If the low level is used to prevent blocking operations, consider logging failed
 
     WIKI_RECOMMENDATION = "Ensure that the return value of a low-level call is checked or logged."
 
-    _txt_description = "low-level calls"
-
     def _is_instance(self, ir):  # pylint: disable=no-self-use
         return isinstance(ir, LowLevelCall)
