+++ slither/detectors/operations/unused_return_values.py
@@ -37,14 +37,14 @@ contract MyConc{
 
     WIKI_RECOMMENDATION = "Ensure that all the return values of the function calls are used."
 
-    _txt_description = "external calls"
-
     def _is_instance(self, ir):  # pylint: disable=no-self-use
-        return (
-            isinstance(ir, HighLevelCall)
-            and isinstance(ir.function, Function)
-            and ir.function.solidity_signature
-            not in ["transfer(address,uint256)", "transferFrom(address,address,uint256)"]
+        return isinstance(ir, HighLevelCall) and (
+            (
+                isinstance(ir.function, Function)
+                and ir.function.solidity_signature
+                not in ["transfer(address,uint256)", "transferFrom(address,address,uint256)"]
+            )
+            or not isinstance(ir.function, Function)
         )
 
     def detect_unused_return_values(self, f):  # pylint: disable=no-self-use
